@@ -14,14 +14,15 @@
   },
   "dependencies": {
     "@tauri-apps/api": "^1.6.0",
-    "bits-ui": "^0.21.16",
+    "bits-ui": "1.0.0-next.45",
     "clsx": "2.1.1",
-    "svelte": "4.2.19",
+    "svelte": "^5.0.0",
     "tailwind-merge": "2.5.3",
     "tailwind-variants": "0.2.1"
   },
   "devDependencies": {
-    "@sveltejs/vite-plugin-svelte": "3.1.2",
+    "@sveltejs/adapter-static": "^3.0.6",
+    "@sveltejs/vite-plugin-svelte": "^4.0.0",
     "@tauri-apps/cli": "^1.6.1",
     "@tsconfig/svelte": "5.0.4",
     "autoprefixer": "^10.4.13",
