@@ -1,10 +1,13 @@
 <script>
   import Login from './Login.svelte';
   import { auth } from '../lib/auth';
+  
+
+  let { children } = $props();
 </script>
 
 {#if $auth.account}
-  <slot />
+  {@render children?.()}
 {:else}
   <!-- user is a guest. Show a login widget instead of the current page. -->
   <Login />
