@@ -4,15 +4,20 @@
 
 	type $$Props = AvatarPrimitive.ImageProps;
 
-	let className: $$Props["class"] = undefined;
-	export let src: $$Props["src"] = undefined;
-	export let alt: $$Props["alt"] = undefined;
-	export { className as class };
+	interface Props {
+		class?: $$Props["class"];
+		src?: $$Props["src"];
+		alt?: $$Props["alt"];
+		[key: string]: any
+	}
+
+	let { class: className = undefined, src = undefined, alt = undefined, ...rest }: Props = $props();
+	
 </script>
 
 <AvatarPrimitive.Image
 	{src}
 	{alt}
 	class={cn("aspect-square h-full w-full", className)}
-	{...$$restProps}
+	{...rest}
 />
