@@ -5,9 +5,14 @@
 	type $$Props = SwitchPrimitive.Props;
 	type $$Events = SwitchPrimitive.Events;
 
-	let className: $$Props["class"] = undefined;
-	export let checked: $$Props["checked"] = undefined;
-	export { className as class };
+	interface Props {
+		class?: $$Props["class"];
+		checked?: $$Props["checked"];
+		[key: string]: any
+	}
+
+	let { class: className = undefined, checked = $bindable(undefined), ...rest }: Props = $props();
+	
 </script>
 
 <SwitchPrimitive.Root
@@ -16,7 +21,7 @@
 		"focus-visible:ring-ring focus-visible:ring-offset-background data-[state=checked]:bg-primary data-[state=unchecked]:bg-input peer inline-flex h-[24px] w-[44px] shrink-0 cursor-pointer items-center rounded-full border-2 border-transparent transition-colors focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-offset-2 disabled:cursor-not-allowed disabled:opacity-50",
 		className
 	)}
-	{...$$restProps}
+	{...rest}
 	on:click
 	on:keydown
 >
