@@ -5,17 +5,22 @@
 
 	type $$Props = DialogPrimitive.OverlayProps;
 
-	let className: $$Props["class"] = undefined;
-	export let transition: $$Props["transition"] = fade;
-	export let transitionConfig: $$Props["transitionConfig"] = {
+	interface Props {
+		class?: $$Props["class"];
+		transition?: $$Props["transition"];
+		transitionConfig?: $$Props["transitionConfig"];
+		[key: string]: any
+	}
+
+	let { class: className = undefined, transition = fade, transitionConfig = {
 		duration: 150,
-	};
-	export { className as class };
+	}, ...rest }: Props = $props();
+	
 </script>
 
 <DialogPrimitive.Overlay
 	{transition}
 	{transitionConfig}
 	class={cn("bg-background/80 fixed inset-0 z-50 backdrop-blur-sm", className)}
-	{...$$restProps}
+	{...rest}
 />
