@@ -4,8 +4,13 @@
 
 	type $$Props = SelectPrimitive.SeparatorProps;
 
-	let className: $$Props["class"] = undefined;
-	export { className as class };
+	interface Props {
+		class?: $$Props["class"];
+		[key: string]: any
+	}
+
+	let { class: className = undefined, ...rest }: Props = $props();
+	
 </script>
 
-<SelectPrimitive.Separator class={cn("bg-muted -mx-1 my-1 h-px", className)} {...$$restProps} />
+<SelectPrimitive.Separator class={cn("bg-muted -mx-1 my-1 h-px", className)} {...rest} />
